requirements.txt
alembic==1.8.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
cachetools==5.2.0
captcha==0.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
google-auth==2.14.1
google-auth-oauthlib==0.7.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
oauthlib==3.2.2
Pillow==9.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.43
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
